arm: tegra: clock: Reading APB bus before disabling clock
It may be possible that write operation on apb bus does not get
complete before disabling clock if the clock is disabled just after
the write on apb bus.
To have proper sequence of operation, it is require to read back the
apb bus to make sure the write operation is completed.

bug 833341

This is cherry pick of change http://git-master/r/#change,32556
in main but hand-merged.

Change-Id: Ifdabd7507d2bd738da3880cff6756a9a7b489e4b
Reviewed-on: http://git-master/r/45981
Reviewed-by: Alok Chauhan <alokc@nvidia.com>
Tested-by: Alok Chauhan <alokc@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>

(krc - also added in 750MHz CPU speed)