arm: tegra: stingray: Enable CMD23 for Toshiba EMMC, remove BFS
Informal testing has shown that there's little (if any?)
performance degredation if we enable CMD23 on the Toshiba EMMC
devices (a possible side-effect of commits 9b4d4a585/d35c2535),
so turn off the quirk, as now configurable in commit 0fb29657.

BFQ scheduler showed the worst performance of the governors tested,
and appears to be optimized for rotational disks anyway, so remove.