mmc: sdhci: fix vmmc handling
Presently the vmmc regulator is enabled when the host
controller is added and disabled when it is removed.
However, the vmmc regulator should be under the control
of the upper layers via ->set_ios().  Make it so.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Chris Ball <cjb@laptop.org>
(cherry picked from commit ceb6143b2df81c8a554bd5de898087926568cc51)

Conflicts:

	drivers/mmc/host/sdhci.c